refactor: reorganize journey ACL imports and remove deprecated file
- Updated import paths for journey ACL in the ability.ts file to reflect new structure.
- Deleted the journey.acl.ts file as it is no longer needed, streamlining the codebase.

This commit improves code organization and reduces clutter by removing unused files.
